remove SignUpInfo & SignUpProfile component
